requirements.txt
lxml>=3.5,<4.0
cssselect>=0.9.1,<1.0
pysolr>=3.3.0,<4.0
OWSLib>=0.10.3,<1.0
demjson>2.0.0,<3.0
unicodecsv==0.14.1
python-dateutil>=2.4.0,<3.0
